Extra Task
Extra Task: Organizing routes in a controller provider
